Final Commit for Project
Includes three csv files that can be used as a demo for this prototype, that can be used by others. Also includes txt files holding past prototypes, and finally a combined prototype of both the Data Analyzer/Feature Extractor & the AI Analyzer/Display so both programs are included in one piece of code.